requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
black==25.12.0
cfgv==3.5.0
click==8.3.1
colorama==0.4.6
distlib==0.4.0
fastapi==0.124.4
filelock==3.20.1
h11==0.16.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
librt==0.7.4
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
ruff==0.14.10
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
virtualenv==20.35.4
